+++ CHANGELOG.rst
@@ -44,6 +44,8 @@ in development
 * Cleanup timers and webhook trigger definitions once all rules referencing them are removed. (bug-fix)
 * Enable pseudo tty when running remote SSH commands with the paramiko SSH runner. This is done
   to match existing Fabric behavior. (bug-fix)
+* Fix CLI so it skips automatic authentication if credentials are provided in the config on "auth"
+  command. (bug fix)
 
 0.13.2 - September 09, 2015
 ---------------------------